+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   - "Limited support:" doesn't attempt to decrypt data and ignores `options:` block
   - allows some detectors to know that keys/data exist, even if it's not known what the data is
 - New subcommand: `driftless report`
+- New subcommand: `driftless list factsets`
 - New detectors `code:unused-modules`, `code:unused-roles`, `code:unused-profiles`:
   modules and role/profile classes that no active node is classified with
 - New config file subsystem `import`: